DAOS-3411 control: record results of each system stop request
SystemStop gRPC from mgmt tool will trigger Stop gRPCs to each system
member (resulting in retried host-local KillRank requests over dRPC).

KillRank will subsequently be replaced by InstanceStop (TBD) requests
to provide graceful shutdown capabilities (incl. inhibition of pool
rebuild).

This change adds SystemMemberResults to SystemStop response.
SystemMemberResults consist of id/errored/message for Stop requests
sent to each system member and enables reporting of individual or
aggregated instance stop request results to user.
