+++ src/control/cmd/dmg/command_test.go
@@ -157,9 +157,9 @@ func (tc *testConn) SystemMemberQuery() (common.SystemMembers, error) {
 	return make(common.SystemMembers, 0), nil
 }
 
-func (tc *testConn) SystemStop() (common.SystemMembers, error) {
+func (tc *testConn) SystemStop() (common.SystemMemberResults, error) {
 	tc.appendInvocation("SystemStop")
-	return make(common.SystemMembers, 0), nil
+	return make(common.SystemMemberResults, 0), nil
 }
 
 func (tc *testConn) SetTransportConfig(cfg *security.TransportConfig) {
